UI updates

+++ pyaudioTest.py
@@ -6,24 +6,47 @@ import time
 import numpy as np
 from scipy.io import wavfile
 import pyaudio
+import collections
 
+BUFFER_SIZE = 1024
+p = pyaudio.PyAudio()
+def record(recording, fs=44100):
+    # Open new audio input stream
+    global stream
+    global buffer
+    buffer = []
+    stream = p.open(format=pyaudio.paInt16, channels=1, rate=fs, input=True, frames_per_buffer=BUFFER_SIZE)
+    #stream.start_stream()
+    # Continuously record audio from the stream and save it to the buffer
+    while True:
+        # Read from the stream
+        data = stream.read(BUFFER_SIZE)
+        # Convert the data to a NumPy array
+        audio_data = np.frombuffer(data, dtype=np.int16)
+        # Add the audio data to the buffer
+        buffer.extend(audio_data)
+        if len(buffer) >= (fs*3) or not recording:
+            stream.stop_stream()
+            stream.close()
+            p.terminate()
+            break
+        
 
-def record(duration=1, fs=44100):
-    nsamples = duration*fs
-    nsamples = int(nsamples)
-    p = pyaudio.PyAudio()
-    stream = p.open(format=pyaudio.paInt16, channels=1, rate=fs, input=True,
-                    frames_per_buffer=nsamples)
-    buffer = stream.read(nsamples)
-    array = np.frombuffer(buffer, dtype='int16')
-    stream.stop_stream()
-    stream.close()
-    p.terminate()
+def stop_record():
+    # Save data
+    array = np.asarray(buffer, dtype='int16')
     return array
 
-
-
+# This method plays back the recoded auio to the user
 def sound(array, fs=44100):
+    """
+    parameters:
+        array -> this array contains the samples for the recorded audio using the buffer above.
+        fs -> this is the sampling rate which is defaulted to 44100 samples per second.
+
+    this function, takes the Pyaudio class, and then opens a stream and writes it onto the buffer, whcih then will be played 
+    through the default Audio player.
+    """
     p = pyaudio.PyAudio()
     stream = p.open(format=pyaudio.paInt16, channels=len(array.shape), rate=fs, output=True)
     stream.write(array.tobytes())
